+++ server/models/Post.js
@@ -49,4 +49,5 @@ const PostSchema = new mongoose.Schema({
     }
 });
 
-module.exports = mongoose.model("Post", PostSchema);
+const Post = mongoose.model("Post", PostSchema);
+module.exports = Post;
